Rename BuildTests.cs to EddiCoreTests.cs, consolidate some relevant tests to that TestClass, and add a keepAlive test.
